Renamed execution-engine-plugin to 'classic-engine-plugin' and created a new bundle for the spark engine where I will merge some POC code in later
