+++ plugins/classic-engine-plugin/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<blueprint xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0"
+           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+           xmlns:pen="http://www.pentaho.com/xml/schemas/pentaho-blueprint"
+           xsi:schemaLocation="
+            http://www.osgi.org/xmlns/blueprint/v1.0.0 http://www.osgi.org/xmlns/blueprint/v1.0.0/blueprint.xsd
+            http://www.pentaho.com/xml/schemas/pentaho-blueprint http://www.pentaho.com/xml/schemas/pentaho-blueprint.xsd">
+
+  <bean id="classicKettleEngine" class="org.pentaho.engine.classic.ClassicKettleEngine" scope="prototype">
+    <pen:di-plugin type="org.pentaho.di.engine.ExecutionEnginePluginType"/>
+  </bean>
+
+</blueprint>
