+++ plugins/pom.xml
@@ -14,6 +14,7 @@
   <modules>
     <module>repositories-plugin</module>
     <module>meta-inject-plugin</module>
-    <module>execution-engine-plugin</module>
+    <module>classic-engine-plugin</module>
+    <module>spark-engine-plugin</module>
   </modules>
 </project>
