+++ plugins/spark-engine-plugin/pom.xml
@@ -7,7 +7,7 @@
     <artifactId>pentaho-ce-jar-parent-pom</artifactId>
     <version>1.2.1</version>
   </parent>
-  <artifactId>pentaho-execution-engine</artifactId>
+  <artifactId>spark-execution-engine</artifactId>
   <version>7.1-SNAPSHOT</version>
   <packaging>bundle</packaging>
   <name>Pentaho Community Edition Project: ${project.artifactId}</name>
