+++ plugins/spark-engine-plugin/res/version.xml
@@ -1,0 +1,2 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<version branch='TRUNK'>@TRUNK@</version>
