+++ plugins/spark-engine-plugin/src/main/java/org/pentaho/engine/spark/SparkEngine.java
@@ -1,14 +1,13 @@
-package org.pentaho.engine.foobar;
+package org.pentaho.engine.spark;
 
 import org.pentaho.di.core.exception.KettleException;
 import org.pentaho.di.engine.ExecutionEngine;
 import org.pentaho.di.engine.IExecutionEngine;
 import org.pentaho.di.trans.Trans;
 
-@ExecutionEngine( id = "FooBarOSGiEngine", name= "Foo Bar OSGi Engine" )
-public class FooBarOSGiEngine implements IExecutionEngine {
+@ExecutionEngine( id = "SparkEngine", name= "Spark Engine" )
+public class SparkEngine implements IExecutionEngine {
 
   @Override public void execute( Trans trans ) throws KettleException {
-
   }
 }
